hello: do not require user 1000
I've tried running the image on a system where rootless users are
using a single ID and it failed.

While it is a good practice to run as a different user, I think that
for such a simple code there is not much gain in running as a
different user, so let's just run as root and avoid completely the
issue.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>